Split up the outputparser tests into multiple files
